[test 164797107]add action tests
