+++ src/api/index.js
@@ -9,6 +9,7 @@ export const setToken = token => {
     destroyToken();
   }
   localStorage.setItem('token', JSON.stringify(token));
+  return getToken();
 };
 
 export const getToken = () => {
@@ -17,5 +18,6 @@ export const getToken = () => {
 };
 
 export const destroyToken = () => {
-  return localStorage.removeItem('token');
+  localStorage.removeItem('token');
+  return null;
 };
