--- src/store/reducers/index.js
@@ -1,5 +1,0 @@
-import { combineReducers } from 'redux';
-
-import auth from './authReducer';
-
-export default combineReducers({ auth });
